# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ocha - Trusted Cleaning & Property Services Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/profile:
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List user properties
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyListResponse"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "201":
          description: Created property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{propertyId}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property profile
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Property profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    put:
      operationId: updateProperty
      tags: [properties]
      summary: Update property profile
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyBody"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /properties/{propertyId}/qr-code:
    get:
      operationId: getPropertyQrCode
      tags: [properties]
      summary: Get QR code URL for property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCodeResponse"
...
  /reviews:
    post:
      operationId: submitReview
      tags: [reviews]
      summary: Submit a review for a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReviewBody"
      responses:
        "201":
          description: Submitted review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/booking/{bookingId}:
    get:
      operationId: getBookingReviews
      tags: [reviews]
      summary: Get reviews for a booking (after reveal)
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews for booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingReviewsResponse"
  /reviews/cleaner/{cleanerId}:
    get:
      operationId: getCleanerReviews
      tags: [reviews]
      summary: Get revealed reviews for a cleaner
      parameters:
        - name: cleanerId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Cleaner reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
...
  /trust/{userId}:
    get:
      operationId: getUserTrustScore
      tags: [trust]
      summary: Get trust score for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trust profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustProfile"
  /trust/verify/phone:
    post:
      operationId: verifyPhone
      tags: [trust]
      summary: Initiate phone verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPhoneBody"
      responses:
        "200":
          description: Verification initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResponse"
  /trust/verify/phone/confirm:
    post:
      operationId: confirmPhoneVerification
      tags: [trust]
      summary: Confirm phone verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPhoneBody"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResponse"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{notificationId}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [customer, cleaner, property_manager, admin]
        isVerified:
          type: boolean
        verificationBadge:
          type: string
          enum: [none, basic, verified, trusted]
        createdAt:
          type: string
        trustScore:
          type: number
      required:
        - id
        - email
        - role
        - isVerified
        - verificationBadge
    UpdateProfileBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
...
    Property:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        propertyType:
          type: string
          enum: [apartment, house, office, airbnb, hmo, serviced_accommodation, other]
        addressLine1:
          type: string
        city:
          type: string
        postcode:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        photoUrl:
          type: string
        entrancePhotoUrl:
          type: string
        accessNotes:
          type: string
        parkingInfo:
          type: string
        hasLift:
          type: boolean
        petsInfo:
          type: string
        cleaningFrequency:
          type: string
          enum: [once, weekly, biweekly, monthly, on_demand]
        preferredCleanerId:
          type: string
        bedroomCount:
          type: integer
        bathroomCount:
          type: integer
        sqft:
          type: integer
        qrCodeUrl:
          type: string
        createdAt:
          type: string
      required:
        - id
        - ownerId
        - name
        - propertyType
        - addressLine1
        - city
        - postcode
    PropertyListResponse:
      type: object
      properties:
        properties:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        total:
          type: integer
      required:
        - properties
        - total
    CreatePropertyBody:
      type: object
      properties:
        name:
          type: string
        propertyType:
          type: string
        addressLine1:
          type: string
        city:
          type: string
        postcode:
          type: string
        accessNotes:
          type: string
        parkingInfo:
          type: string
        hasLift:
          type: boolean
        petsInfo:
          type: string
        cleaningFrequency:
          type: string
        bedroomCount:
          type: integer
        bathroomCount:
          type: integer
        sqft:
          type: integer
      required:
        - name
        - propertyType
        - addressLine1
        - city
        - postcode
    UpdatePropertyBody:
      type: object
      properties:
        name:
          type: string
        accessNotes:
          type: string
        parkingInfo:
          type: string
        hasLift:
          type: boolean
        petsInfo:
          type: string
        cleaningFrequency:
          type: string
        preferredCleanerId:
          type: string
        photoUrl:
          type: string
    QrCodeResponse:
      type: object
      properties:
        qrCodeUrl:
          type: string
        deepLink:
          type: string
        propertyId:
          type: string
      required:
        - qrCodeUrl
        - deepLink
        - propertyId
...
    Review:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        reviewerId:
          type: string
        revieweeId:
          type: string
        reviewerRole:
          type: string
          enum: [customer, cleaner]
        punctualityRating:
          type: integer
        professionalismRating:
          type: integer
        qualityRating:
          type: integer
        communicationRating:
          type: integer
        reliabilityRating:
          type: integer
        overallRating:
          type: number
        comment:
          type: string
        wouldWorkAgain:
          type: boolean
        isRevealed:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - bookingId
        - reviewerId
        - revieweeId
        - reviewerRole
        - overallRating
        - wouldWorkAgain
        - isRevealed
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        averageRating:
          type: number
        wouldWorkAgainPct:
          type: number
      required:
        - reviews
        - total
    BookingReviewsResponse:
      type: object
      properties:
        bookingId:
          type: string
        customerReview:
          $ref: "#/components/schemas/Review"
        cleanerReview:
          $ref: "#/components/schemas/Review"
        areRevealed:
          type: boolean
        revealAt:
          type: string
      required:
        - bookingId
        - areRevealed
    SubmitReviewBody:
      type: object
      properties:
        bookingId:
          type: string
        punctualityRating:
          type: integer
        professionalismRating:
          type: integer
        qualityRating:
          type: integer
        communicationRating:
          type: integer
        reliabilityRating:
          type: integer
        comment:
          type: string
        wouldWorkAgain:
          type: boolean
      required:
        - bookingId
        - wouldWorkAgain
    TrustProfile:
      type: object
      properties:
        userId:
          type: string
        overallScore:
          type: number
        verificationStatus:
          $ref: "#/components/schemas/VerificationStatus"
        behavioralMetrics:
          $ref: "#/components/schemas/BehavioralMetrics"
        badge:
          type: string
          enum: [none, basic, verified, trusted]
      required:
        - userId
        - overallScore
        - verificationStatus
        - badge
    VerificationStatus:
      type: object
      properties:
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        idVerified:
          type: boolean
        selfieVerified:
          type: boolean
        addressVerified:
          type: boolean
        backgroundCheckDone:
          type: boolean
        paymentVerified:
          type: boolean
      required:
        - emailVerified
        - phoneVerified
        - idVerified
        - selfieVerified
        - addressVerified
        - backgroundCheckDone
        - paymentVerified
    BehavioralMetrics:
      type: object
      properties:
        cancellationRate:
          type: number
        repeatBookingRate:
          type: number
        responseSpeed:
          type: string
        latePct:
          type: number
        disputeFrequency:
          type: number
        wouldWorkAgainPct:
          type: number
        completionRate:
          type: number
      required:
        - cancellationRate
        - repeatBookingRate
        - completionRate
    VerifyPhoneBody:
      type: object
      properties:
        phoneNumber:
          type: string
      required:
        - phoneNumber
    ConfirmPhoneBody:
      type: object
      properties:
        phoneNumber:
          type: string
        code:
          type: string
      required:
        - phoneNumber
        - code
    VerificationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        verified:
          type: boolean
      required:
        - success
...
    Location:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
      required:
        - lat
        - lng
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [booking_request, booking_accepted, booking_en_route, booking_started, booking_completed, review_request, review_revealed, trust_update, system]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        bookingId:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - title
        - message
        - isRead
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        unreadCount:
          type: integer
      required:
        - notifications
        - total
        - unreadCount
...